+++ src/main/java/baseball/PlayBall.java
@@ -1,0 +1,16 @@
+package baseball;
+
+import baseball.number.ComputerNumber;
+
+public class PlayBall {
+    ComputerNumber computerNumber;
+
+    public void startGame() {
+        computerNumber = new ComputerNumber();
+        computerNumber.creatComputerNumber();
+    }
+
+    private void getUserInput(){
+
+    }
+}