+++ CHANGELOG.md
@@ -7,12 +7,21 @@
 1. ```
    2021/11/24
    "feat : add README.md to list up application functions"
+   
    프로젝트 생성 및 요구사항에 맞는 기능 목록을 포함하여 리드미 작성
    ```
 
    
 
+2. ```
+   2021/11/24
+   "feat : add creating random computer number with handling exception"
+   
+   pickNumberInRange()를 활용하여 1 ~ 9 사이의 서로 다른 세 자리 정수를 생성하는 기능 추가
+   랜덤하게 생성된 숫자의 예외처리 기능(중복) 추가
+   ```
 
+